backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.150.0
apscheduler==3.10.4
python-dotenv==1.0.1
anthropic==0.36.1
